--story=887060303 环境管理服务对接上云版作业平台
prefer：优化变量类型

+++ src/backend/ci/core/environment/biz-environment/src/main/kotlin/com/tencent/devops/environment/service/job/QueryAgentStatusService.kt
@@ -82,7 +82,7 @@ class QueryAgentStatusService @Autowired constructor(
     private fun getAgentVersionsFromJob(
         userId: String,
         projectId: String,
-        hostIdList: List<Int>
+        hostIdList: List<Long>
     ): AgentResult<QueryAgentStatusFromJobResult> {
         val queryAgentStatusRequest = QueryAgentStatusFromJobReq(
             hostIdList = hostIdList
